requirements.txt
pipecat-ai[cartesia,openai,silero,deepgram,assemblyai,cerebras]
fastapi
uvicorn
python-dotenv
loguru
